world-map-renderer/requirements.txt
Pillow
confluent-kafka